requirements.txt
jsonschema==3.2.0
matplotlib==3.8.2
pandas==2.1.3
Pillow==9.0.1
protobuf==3.20.1
scikit_image==0.18.3
scipy==1.7.1
tensorboard==2.7.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.66.1
pydantic==2.5.2
PyYAML==6.0.1